+++ codegen/src/main/kotlin/io/github/enjoydambience/kotlinbard/codegen/generators/BuilderTypeAliases.kt
@@ -14,13 +14,13 @@
  *    limitations under the License.
  */
 
-package io.github.enjoydambience.kotlinbard.codegen.generation
+package io.github.enjoydambience.kotlinbard.codegen.generators
 
-import com.squareup.kotlinpoet.AnnotationSpec
-import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
+import io.github.enjoydambience.kotlinbard.CodegenDsl
 import io.github.enjoydambience.kotlinbard.addTypeAlias
-import io.github.enjoydambience.kotlinbard.codegen.destinationPackage
+import io.github.enjoydambience.kotlinbard.buildAnnotation
+import io.github.enjoydambience.kotlinbard.codegen.SpecInfo
 
 /**
  * Creates typeAliases for all spec builders.
@@ -28,9 +28,7 @@ import io.github.enjoydambience.kotlinbard.codegen.destinationPackage
  * These aliases are used so builders can have a DSL annotation.
  */
 object BuilderTypeAliases : FileGenerator {
-    private val dslAnnotation = AnnotationSpec.builder(ClassName(destinationPackage, "CodegenDsl")).build()
-
-    override val fileName: String get() = "_BuilderTypeAliases"
+    private val dslAnnotation = buildAnnotation(CodegenDsl::class)
 
     override fun FileSpec.Builder.generate() {
         SpecInfo.allSpecs.forEach {