+++ codegen/src/main/kotlin/io/github/enjoydambience/kotlinbard/codegen/generators/SpecBuilders.kt
@@ -14,29 +14,30 @@
  *    limitations under the License.
  */
 
-package io.github.enjoydambience.kotlinbard.codegen.generation
+package io.github.enjoydambience.kotlinbard.codegen.generators
 
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.LambdaTypeName
 import com.squareup.kotlinpoet.UNIT
 import io.github.enjoydambience.kotlinbard.addParameter
+import io.github.enjoydambience.kotlinbard.buildFunction
+import io.github.enjoydambience.kotlinbard.codegen.SpecInfo
 import io.github.enjoydambience.kotlinbard.codegen.copyDeprecationOf
 import io.github.enjoydambience.kotlinbard.codegen.reflectCodeCall
-import io.github.enjoydambience.kotlinbard.createFunction
 import net.pearx.kasechange.toPascalCase
 import kotlin.reflect.KFunction
 import kotlin.reflect.full.declaredMemberFunctions
 
 /**
- * Generates spec creator functions (`createXXX` functions).
+ * Generates spec creator functions (`buildXXX` functions).
  * These are called "creators" instead of "builders" to disambiguate from kotlin-poet's spec builders.
  *
  * These are derived from spec companion functions that return a builder type.
  *
- * These functions have the form `XXXSpec.createXxx(<parameters>).apply(config).build()`.
+ * These functions have the form `XXXSpec.xxx(<parameters>).apply(config).build()`.
  */
-object SpecBuilders : SpecFunctionFileGenerator("_SpecBuilders") {
+object SpecBuilders : SpecFunctionFileGenerator() {
     const val funPrefix = "build"
     override fun generateFunctionsForSpec(spec: SpecInfo): List<FunSpec> =
         spec.companionClass.declaredMemberFunctions
@@ -48,7 +49,7 @@ object SpecBuilders : SpecFunctionFileGenerator("_SpecBuilders") {
             }
 
     private fun generateFunction(spec: SpecInfo, function: KFunction<*>): FunSpec {
-        // "xxxBuilder" -> "createXxx"
+        // "xxxBuilder" -> "buildXxx"
         // "builder" -> "create<spec name>"
         val funNameMinusBuilder = function.name.replace("[bB]uilder".toRegex(), "")
         val generatedName = if (spec.name == "Type" && funNameMinusBuilder == "annotation") {
@@ -60,7 +61,7 @@ object SpecBuilders : SpecFunctionFileGenerator("_SpecBuilders") {
             funPrefix + it.toPascalCase()
         }
 
-        return createFunction(generatedName) {
+        return buildFunction(generatedName) {
             copyDeprecationOf(function)
 
             addModifiers(KModifier.INLINE)