+++ Changelog.md
@@ -1,5 +1,11 @@
-# Version 0.0.2
-- Added control flow dsl
+# Version 0.2.0
+- Added extension shortcut CodeBlock creators
+- Added parameters dsl
+- Added SpecModifiers
+- Added more add extensions
+- Remove need to re-specify type in property setter
+
+# Version 0.1.0
 - Disambiguate `annotation` and `annotationClass`
 - DslMarker for builder scopes
 - Empty setter
@@ -7,6 +13,9 @@
 - `init` blocks
 - `addConstructor` for secondary constructors
 - `addEnumConstant`
+
+# Version 0.0.2
+- Added control flow dsl
 
 # Version 0.0.1
 - Initial release